Модуль 33. Рефакторинг приложение в соответствии с паттерном архитектуры приложений MVVM.
