+++ app/src/main/java/ru/seraf1n/moviefinder/domain/Interactor.kt
@@ -1,0 +1,7 @@
+package ru.seraf1n.moviefinder.domain
+
+import ru.seraf1n.moviefinder.data.MainRepository
+
+class Interactor(val repo: MainRepository) {
+    fun getFilmsDB(): List<Film> = repo.filmsDataBase
+}
