+++ app/src/main/java/ru/seraf1n/moviefinder/view/SplashActivity.kt
@@ -1,4 +1,4 @@
-package ru.seraf1n.moviefinder
+package ru.seraf1n.moviefinder.view
 
 import android.content.Intent
 import android.os.Bundle
@@ -7,8 +7,12 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
+import ru.seraf1n.moviefinder.R
+
+const val STANDART_DELAY = 3000L
 
 class SplashActivity : AppCompatActivity() {
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_splash)
@@ -17,7 +21,7 @@ class SplashActivity : AppCompatActivity() {
         CoroutineScope(Dispatchers.Main).launch {
 
             val intent = Intent(this@SplashActivity, MainActivity::class.java)
-            delay(3000)
+            delay(STANDART_DELAY)
             startActivity(intent)
             finish()
