+++ app/src/main/java/ru/seraf1n/moviefinder/view/fragments/FavoritesFragment.kt
@@ -1,4 +1,4 @@
-package ru.seraf1n.moviefinder
+package ru.seraf1n.moviefinder.view.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,7 +7,11 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearLayoutManager
 import kotlinx.android.synthetic.main.fragment_favorities.view.*
+import ru.seraf1n.moviefinder.utils.AnimationHelper
+import ru.seraf1n.moviefinder.view.rv_adapters.FilmListRecyclerAdapter
+import ru.seraf1n.moviefinder.view.MainActivity
 import ru.seraf1n.moviefinder.databinding.FragmentFavoritiesBinding
+import ru.seraf1n.moviefinder.domain.Film
 
 class FavoritesFragment : Fragment() {
 
@@ -25,7 +29,11 @@ class FavoritesFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        AnimationHelper.performFragmentCircularRevealAnimation(binding.fragmentFavoritiesRoot, requireActivity(), 1)
+        AnimationHelper.performFragmentCircularRevealAnimation(
+            binding.fragmentFavoritiesRoot,
+            requireActivity(),
+            1
+        )
         //Получаем список при транзакции фрагмента
         val favoritesList: List<Film> = emptyList()
         binding.root.favorites_recycler
