+++ app/src/main/java/ru/seraf1n/moviefinder/view/rv_viewholders/FilmViewHolder.kt
@@ -1,9 +1,10 @@
-package ru.seraf1n.moviefinder
+package ru.seraf1n.moviefinder.view.rv_viewholders
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import kotlinx.android.synthetic.main.film_item.view.*
+import ru.seraf1n.moviefinder.domain.Film
 
 const val RATING_INDEX = 10
 //В конструктор класс передается layout, который мы создали(film_item.xml)
