+++ app/src/main/java/ru/seraf1n/moviefinder/viewmodel/HomeFragmentViewModel.kt
@@ -1,0 +1,18 @@
+package ru.seraf1n.moviefinder.viewmodel
+
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import ru.seraf1n.moviefinder.App
+import ru.seraf1n.moviefinder.domain.Film
+import ru.seraf1n.moviefinder.domain.Interactor
+
+class HomeFragmentViewModel : ViewModel() {
+    val filmsListLiveData:  MutableLiveData<List<Film>> = MutableLiveData()
+    //Инициализируем интерактор
+    private var interactor: Interactor = App.instance.interactor
+
+    init {
+        val films = interactor.getFilmsDB()
+        filmsListLiveData.postValue(films)
+    }
+}
