+++ app/src/main/res/layout/activity_splash.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SplashActivity">
+    tools:context=".view.SplashActivity">
     <com.airbnb.lottie.LottieAnimationView
         android:id="@+id/lottie_anim"
         android:layout_width="@dimen/splash_width"
