+++ app/src/main/res/layout/fragment_details.xml
@@ -6,7 +6,7 @@
     android:fitsSystemWindows="true"
     android:id="@+id/fragment_details"
     android:background="@drawable/corner_background"
-    tools:context=".MainActivity">
+    tools:context=".view.MainActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/app_bar"
