+++ app/src/main/res/layout/fragment_selections.xml
@@ -5,6 +5,6 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/corner_background_collections"
-    tools:context=".SelectionsFragment">
+    tools:context=".view.fragments.SelectionsFragment">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
