+++ app/src/main/res/layout/fragment_watch_later.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/corner_background_later"
-    tools:context=".WatchLaterFragment">
+    tools:context=".view.fragments.WatchLaterFragment">
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
